Add NodeMCUController
Start for the API to convert the NodeMCU Values

Created the NodeMCUCOntroller and NodeMCU class for serializing the requests into json.  Using the json.net from Newtonsoft and had to delete one entry in the .csproj file because there was a double entry.
the Controller will recieve values and add it to a json file which will be used to send these to the Android app.